spot check ladders
