+++ src/data/zip_spotcheck/spot_check_zip_summaries.csv
@@ -1,0 +1,159 @@
+zip,avg_miles_from_center,median_miles_from_center,count_locations
+81501,1.2,1.3,11
+80033,1.1,1.3,7
+80112,7.6,6.7,7
+80524,2.5,2.1,7
+80022,5.2,5.6,6
+80120,1.6,1.7,6
+80205,6.0,6.2,6
+80206,4.5,5.0,6
+80501,0.9,0.8,6
+80525,1.4,1.6,6
+81623,1.7,0.8,6
+80003,2.5,1.7,5
+80204,7.9,8.3,5
+81001,2.3,2.2,5
+81416,0.9,0.4,5
+80215,12.6,12.3,4
+80219,9.4,9.3,4
+80236,12.2,12.2,4
+80631,2.7,2.7,4
+80918,4.1,4.1,4
+81101,0.9,0.8,4
+80010,5.0,5.1,3
+80021,3.5,3.1,3
+80030,3.2,3.4,3
+80045,4.1,4.2,3
+80110,1.3,1.1,3
+80128,5.3,4.7,3
+80134,2.0,2.0,3
+80203,6.3,6.4,3
+80216,5.1,6.3,3
+80222,6.8,6.9,3
+80223,7.0,7.0,3
+80227,13.0,13.0,3
+80229,8.9,9.1,3
+80231,6.6,7.3,3
+80302,1.6,1.4,3
+80303,0.8,0.9,3
+80304,1.6,1.8,3
+80401,1.2,0.9,3
+80439,4.2,3.2,3
+80526,1.9,1.9,3
+80601,2.8,3.0,3
+80634,0.8,0.9,3
+80751,1.3,1.3,3
+80909,2.5,2.4,3
+81050,0.8,0.8,3
+81301,0.6,0.5,3
+81401,0.9,0.8,3
+81601,1.3,0.6,3
+81611,1.0,0.9,3
+80011,2.2,2.2,2
+80014,3.2,3.2,2
+80026,0.3,0.3,2
+80027,1.5,1.5,2
+80111,5.8,5.8,2
+80138,1.7,1.7,2
+80210,7.5,7.5,2
+80220,3.7,3.7,2
+80221,9.4,9.4,2
+80228,14.6,14.6,2
+80234,12.4,12.4,2
+80239,2.8,2.8,2
+80260,9.6,9.6,2
+80517,1.4,1.4,2
+80537,1.1,1.1,2
+80538,1.2,1.2,2
+80701,0.9,0.9,2
+80758,1.1,1.1,2
+80807,0.6,0.6,2
+80810,0.8,0.8,2
+80904,5.4,5.4,2
+81003,1.1,1.1,2
+81052,0.8,0.8,2
+81212,1.8,1.8,2
+81224,0.1,0.1,2
+81230,0.7,0.7,2
+81321,0.6,0.6,2
+81625,0.5,0.5,2
+81632,1.2,1.2,2
+81657,0.5,0.5,2
+80012,3.2,3.2,1
+80031,2.7,2.7,1
+80040,4.2,4.2,1
+80104,0.8,0.8,1
+80107,0.4,0.4,1
+80108,2.6,2.6,1
+80113,1.5,1.5,1
+80121,1.8,1.8,1
+80124,7.8,7.8,1
+80130,1.9,1.9,1
+80201,6.2,6.2,1
+80202,6.8,6.8,1
+80208,7.8,7.8,1
+80211,7.0,7.0,1
+80214,10.5,10.5,1
+80218,5.5,5.5,1
+80226,10.7,10.7,1
+80230,2.5,2.5,1
+80233,11.5,11.5,1
+80235,13.4,13.4,1
+80243,11.2,11.2,1
+80246,5.4,5.4,1
+80305,2.3,2.3,1
+80422,2.6,2.6,1
+80424,0.7,0.7,1
+80433,0.5,0.5,1
+80444,0.5,0.5,1
+80446,0.6,0.6,1
+80452,7.1,7.1,1
+80461,0.6,0.6,1
+80477,1.7,1.7,1
+80481,4.9,4.9,1
+80487,1.4,1.4,1
+80504,3.7,3.7,1
+80514,1.0,1.0,1
+80530,0.2,0.2,1
+80549,1.0,1.0,1
+80640,6.2,6.2,1
+80720,0.5,0.5,1
+80734,0.4,0.4,1
+80737,0.6,0.6,1
+80759,0.4,0.4,1
+80809,0.7,0.7,1
+80817,1.7,1.7,1
+80828,0.3,0.3,1
+80863,0.5,0.5,1
+80905,4.2,4.2,1
+80906,5.2,5.2,1
+80910,2.5,2.5,1
+80923,5.9,5.9,1
+81004,0.7,0.7,1
+81005,3.2,3.2,1
+81054,0.5,0.5,1
+81063,6.0,6.0,1
+81067,1.0,1.0,1
+81073,0.5,0.5,1
+81082,0.8,0.8,1
+81089,0.4,0.4,1
+81120,0.7,0.7,1
+81125,0.3,0.3,1
+81144,0.9,0.9,1
+81151,3.7,3.7,1
+81152,0.2,0.2,1
+81201,0.4,0.4,1
+81423,0.3,0.3,1
+81425,1.6,1.6,1
+81428,3.0,3.0,1
+81432,0.1,0.1,1
+81435,5.0,5.0,1
+81503,2.4,2.4,1
+81504,2.9,2.9,1
+81520,1.6,1.6,1
+81521,0.8,0.8,1
+81602,0.2,0.2,1
+81620,0.9,0.9,1
+81631,0.5,0.5,1
+81641,1.5,1.5,1
+81650,0.7,0.7,1
